docs(kiarina-lib-falkordb): update CHANGELOG for module structure refactoring
