+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - **kiarina-lib-cloudflare-d1**: **BREAKING** - Refactored internal module structure and renamed function parameters for consistency
+- **kiarina-lib-falkordb**: Refactored internal module structure following project architecture rules
 
 ## [1.11.2] - 2025-12-02
