+++ packages/kiarina-lib-falkordb/src/kiarina/lib/falkordb/_core/helpers/get_falkordb.py
@@ -7,7 +7,7 @@ import redis.asyncio
 from redis.backoff import ExponentialBackoff
 from redis.retry import Retry
 
-from ..settings import settings_manager
+from ..._settings import settings_manager
 
 _sync_cache: dict[str, falkordb.FalkorDB] = {}
 """
