+++ packages/kiarina-lib-falkordb/src/kiarina/lib/falkordb/asyncio.py
@@ -3,10 +3,16 @@ from importlib import import_module
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from ._async.registry import get_falkordb
-    from .settings import settings_manager
+    from ._async.helpers.get_falkordb import get_falkordb
+    from ._settings import FalkorDBSettings, settings_manager
 
-__all__ = ["get_falkordb", "settings_manager"]
+__all__ = [
+    # ._async._helpers
+    "get_falkordb",
+    # ._settings
+    "FalkorDBSettings",
+    "settings_manager",
+]
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
@@ -16,8 +22,11 @@ def __getattr__(name: str) -> object:
         raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
 
     module_map = {
-        "get_falkordb": "._async.registry",
-        "settings_manager": ".settings",
+        # ._async.helpers
+        "get_falkordb": "._async.helpers.get_falkordb",
+        # ._settings
+        "FalkorDBSettings": "._settings",
+        "settings_manager": "._settings",
     }
 
     parent = __name__.rsplit(".", 1)[0]
